+++ sdmx-dl-util/src/main/java/sdmxdl/util/BaseProperty.java
@@ -3,7 +3,11 @@ package sdmxdl.util;
 import nbbrd.design.SealedType;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 @SealedType({Property.class, BooleanProperty.class, IntProperty.class, LongProperty.class})
 public abstract class BaseProperty implements CharSequence {
@@ -38,5 +42,11 @@ public abstract class BaseProperty implements CharSequence {
     @Override
     public String toString() {
         return getKey();
+    }
+
+    public static List<String> keysOf(BaseProperty... properties) {
+        return Stream.of(properties)
+                .map(BaseProperty::getKey)
+                .collect(Collectors.collectingAndThen(Collectors.toList(), Collections::unmodifiableList));
     }
 }